auth-proxy-sidecar/requirements-simple.txt
azure-core==1.35.0
azure-identity==1.24.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==45.0.7
flask==3.1.2
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
pycparser==2.22 ; platform_python_implementation != 'PyPy'
pyjwt==2.10.1
requests==2.32.5
six==1.17.0
typing-extensions==4.15.0
urllib3==2.5.0
werkzeug==3.1.3
